update map and improve bot logic
