+++ maps/fortress.txt
@@ -1,6 +1,6 @@
 -----------------------
----------------------2-
 -----------------------
+-------------------2---
 TTTTTTTTTTTTTTTTTTTTTTT
 TTTTTTTTTTTTTTTTTTTTTTT
 TTTTTTTTTTTTTTTTTTTTTTT
@@ -18,6 +18,6 @@ TTTTTTTTTTTTTTTTTTTTTTT
 TTTTTTTTTTTTTTTTTTTTTTT
 TTTTTTTTTTTTTTTTTTTTTTT
 TTTTTTTTTTTTTTTTTTTTTTT
+---1-------------------
 -----------------------
--1---------------------
 -----------------------
