+++ maps/maze.txt
@@ -1,6 +1,6 @@
 ----B------------------
-----B----------------2-
 ----B------------------
+----B--------------2---
 BBBBSSSSSTTTTTSSSSS----
 ----S-------------S----
 ----S-------------S----
@@ -18,6 +18,6 @@ BBBBSSSSSTTTTTSSSSS----
 ----S-------------S----
 ----S-------------S----
 ----SSSSSTTTTTSSSSSBBBB
+---1--------------B----
 ------------------B----
--1----------------B----
 ------------------B----
