+++ maps/plain.txt
@@ -1,5 +1,6 @@
 -----------------------
 -----------------------
+-------------------2---
 -----------------------
 -----------------------
 -----------------------
@@ -9,7 +10,6 @@
 -----------------------
 -----------------------
 -----------------------
------------2-----------
 -----------------------
 -----------------------
 -----------------------
@@ -18,6 +18,6 @@
 -----------------------
 -----------------------
 -----------------------
+---1-------------------
 -----------------------
--1---------------------
 -----------------------
