+++ src/wobject/bot/strategy/IdiotStrategy.java
@@ -3,6 +3,8 @@ package wobject.bot.strategy;
 import wobject.World;
 import wobject.command.GameCommand;
 
+import java.util.Random;
+
 import static wobject.command.CommandFactory.getRandomMove;
 
 public class IdiotStrategy extends Strategy {
@@ -13,6 +15,10 @@ public class IdiotStrategy extends Strategy {
 
     @Override
     public GameCommand getNextMove() {
-        return getRandomMove(world, botNumber);
+        if(new Random().nextInt() % 3 == 0) {
+            return getRandomMove(world, botNumber);
+        } else {
+            return new StrategyUtils(world, botNumber).moveToObject(world.getTanks().get(0));
+        }
     }
 }
